Still pretty bs implementation, but remembers values across lines now
I played around with the vm object quite a bit, cannot fkn figure out how to use the debug thing.
Pretty sure it can get the values I want since the Chrome Devtools can do this,
but I can't figure out how to like give it code and get a callback for each line,
or set breakpoints for each line (you have to give it the code ahead of time and it doesn't have
an interface that will let you pass breakpoints at that point. TBH, I don't really see how it's
different from the regular vm methods).

Anyway, props to v8 for being well enough written to allow things like this,
it's a flaw in Ruby that it doesn't support such things.